# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vullify API
  description: |
    REST API for container registry integration, vulnerability scans (Trivy), findings,
    enrichment (EPSS / KEV), and reporting. List responses wrap payloads in `data` with optional `meta` pagination.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/registries:
    get:
      tags: [Registries]
      summary: List registries
      operationId: listRegistries
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Paginated list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistryListEnvelope"
    post:
      tags: [Registries]
      summary: Create registry
      description: Validates connector credentials with a live connection test before insert.
      operationId: createRegistry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistryCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistryEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "422":
          description: Registry connection failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/registries/{id}:
    parameters:
      - $ref: "#/components/parameters/RegistryId"
    get:
      tags: [Registries]
      summary: Get registry
      operationId: getRegistry
      responses:
        "200":
          description: Registry detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistryEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Registries]
      summary: Update registry
      operationId: updateRegistry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistryEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Registry connection failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    delete:
      tags: [Registries]
      summary: Soft-delete registry
      operationId: deleteRegistry
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/registries/{id}/sync:
    parameters:
      - $ref: "#/components/parameters/RegistryId"
    post:
      tags: [Registries]
      summary: Sync images from registry
      description: Lists repositories and tags from the registry and upserts image rows.
      operationId: syncRegistry
      responses:
        "200":
          description: Sync summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      synced_images:
                        type: integer
                      repositories_seen:
                        type: integer
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Registry sync failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...
components:
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    RegistryId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error:
              code: VALIDATION_ERROR
              message: invalid JSON body
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    Meta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
          format: int64
    Envelope:
      type: object
      properties:
        data: {}
        meta:
          $ref: "#/components/schemas/Meta"
    Registry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          description: dockerhub, gitlab, ecr, gcr, ghcr
        url:
          type: string
        credentials_set:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RegistryCreate:
      type: object
      required: [name, type, url]
      properties:
        name:
          type: string
        type:
          type: string
        url:
          type: string
        credentials:
          description: Connector-specific JSON (default {})
          type: object
    RegistryUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        url:
          type: string
        credentials:
          type: object
    RegistryEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Registry"
    RegistryListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Registry"
...